Refactor file.py
